requirements.txt
1) Instalar dependencias de la lista a continuacion mediante pip install

blinker==1.7.0
click==8.1.7   
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
python-dotenv==1.0.0
Werkzeug==3.0.1

2)Crear la base de datos para poder trabajar localmente (ejecutar script schema.sql)

3)Generar un archivo que se llame '.env' en la raíz del proyecto con las siguientes variables:
DB_HOST='host de la DB' en general es localhost
DB_USER='usuario de la db' por defecto es root
DB_PASSWORD='password de la db' password que hayan elegido, en caso de no tener password avisarme y modificamos dos lineas de codigo
DB_DATABASE='codoacodo' es el nombre que le di a la db en el script.
DB_PORT='puerto en el que este la db' por defecto es el 3306.

Este archivo es para que cada uno pueda poner sus credenciales de acceso a la base de datos y no tener que estar modificando el codigo fuente.

4)

Los endpoints de la api de desarrollo son

GET     http://127.0.0.1:5000/cursos          -- Para listar todos los cursos
POST    http://127.0.0.1:5000/cursos          -- Para crear un nuevo cursos
PUT     http://127.0.0.1:5000/cursos/idcurso  -- Para modificar un curso existente (reemplazar idcurso por numero de id en la db)
DELETE  http://127.0.0.1:5000/cursos/idcurso  -- Para eliminar un curso existente (reemplazar idcurso por numero de id en la db)


